refactor(chat): extract inventory-tree-source; record migration conventions
Phase-1 step 1 of the inventory-bar decomposition: moves the
substrate-agnostic data layer out of the 1267-line inventoryComponent into
inventory-tree-source.js — the formula-type classification constants
(CONVERSABLE_TYPES, NON_EXPANDABLE_TYPES, HUB_TYPES) and the static-snapshot
adapter (makeStaticNameIterator, makeStaticTreePowers). A verbatim move,
re-imported at the original call sites; this code survives the eventual
Preact conversion unchanged.

Also records two migration conventions in the design doc:
- No JSX. Components are plain .js authored with the h() helper from
  setup-preact-container.js (matching @endo/preact-container's own suite).
- Two-phase rule: extract substrate-agnostic logic/behavior/boundaries in
  the current imperative style first (survives conversion); rewrite markup
  directly in h() rather than building throwaway imperative plumbing. The
  inventory pieces are classified phase-1 vs convert-in-place accordingly.

Verified behavior-preserving: node --check, eslint (incl. import
resolution), and prettier all clean; tsc error count unchanged vs HEAD
(9 pre-existing Passable/iterateReader errors, none introduced).
